+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.3.38 Release:
 ## New Generic Monster Features:
+- Activity On Effect Expiry
 - Blood Frenzy
 - Bloodied Frenzy
 - Escape